Convert vnet_config to handler format (#60370)
[Client Tools Updates](https://github.com/gravitational/teleport/blob/master/rfd/0144-client-tools-updates.md)
is disabled to ensure the locally-compiled binary is used.

```bash
$ export TELEPORT_TOOLS_VERSION=off

$ ./tctl get vnet_config
kind: vnet_config
metadata:
  name: vnet-config
  revision: f2:30b99fcf-d2ac-442d-96d0-93fbf747f707
spec:
  ipv4_cidr_range: 100.64.0.0/10
version: v1

$ ./tctl edit vnet_config
ERROR: rpc error: code = Unknown desc = parsing ipv4_cidr_range
        invalid CIDR address: 198.18.0/15

$ ./tctl edit vnet_config
vnet_config has been updated

$ ./tctl get vnet_config
kind: vnet_config
metadata:
  name: vnet-config
  revision: f2:e20be0d6-290f-425e-bd45-722cfaa76db5
spec:
  ipv4_cidr_range: 198.18.0.0/15
version: v1

$ ./tctl rm vnet_config
vnet_config has been deleted

$ ./tctl get vnet_config
ERROR: getting resource "" of type "vnet_config"
        vnet_config "vnet-config" doesn't exist

```
